backend/requirements.txt
flask
flask-cors
python-dotenv
openai
flask-sqlalchemy
requests
